+++ app/src/main/java/com/naman14/timber/activities/SearchActivity.java
@@ -15,7 +15,9 @@
 package com.naman14.timber.activities;
 
 import android.content.Context;
+import android.os.AsyncTask;
 import android.os.Bundle;
+import android.support.annotation.Nullable;
 import android.support.v4.view.MenuItemCompat;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -38,11 +40,16 @@ import com.naman14.timber.models.Song;
 import com.naman14.timber.provider.SearchHistory;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
 
 public class SearchActivity extends BaseThemedActivity implements SearchView.OnQueryTextListener, View.OnTouchListener {
+
+    private final Executor mSearchExecutor = Executors.newSingleThreadExecutor();
+    @Nullable
+    private AsyncTask mSearchTask = null;
 
     private SearchView mSearchView;
     private InputMethodManager mImm;
@@ -51,7 +58,7 @@ public class SearchActivity extends BaseThemedActivity implements SearchView.OnQ
     private SearchAdapter adapter;
     private RecyclerView recyclerView;
 
-    private List searchResults = Collections.emptyList();
+    private List<Object> searchResults = Collections.emptyList();
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -128,33 +135,18 @@ public class SearchActivity extends BaseThemedActivity implements SearchView.OnQ
         if (newText.equals(queryString)) {
             return true;
         }
+        if (mSearchTask != null) {
+            mSearchTask.cancel(false);
+            mSearchTask = null;
+        }
         queryString = newText;
-        if (!queryString.trim().equals("")) {
-            this.searchResults = new ArrayList();
-            List<Song> songList = SongLoader.searchSongs(this, queryString);
-            List<Album> albumList = AlbumLoader.getAlbums(this, queryString);
-            List<Artist> artistList = ArtistLoader.getArtists(this, queryString);
-
-            if (!songList.isEmpty()) {
-                searchResults.add("Songs");
-            }
-            searchResults.addAll((Collection) (songList.size() < 10 ? songList : songList.subList(0, 10)));
-            if (!albumList.isEmpty()) {
-                searchResults.add("Albums");
-            }
-            searchResults.addAll((Collection) (albumList.size() < 7 ? albumList : albumList.subList(0, 7)));
-            if (!artistList.isEmpty()) {
-                searchResults.add("Artists");
-            }
-            searchResults.addAll((Collection) (artistList.size() < 7 ? artistList : artistList.subList(0, 7)));
-        } else {
+        if (queryString.trim().equals("")) {
             searchResults.clear();
             adapter.updateSearchResults(searchResults);
             adapter.notifyDataSetChanged();
+        } else {
+            mSearchTask = new SearchTask().executeOnExecutor(mSearchExecutor, queryString);
         }
-
-        adapter.updateSearchResults(searchResults);
-        adapter.notifyDataSetChanged();
 
         return true;
     }
@@ -165,6 +157,14 @@ public class SearchActivity extends BaseThemedActivity implements SearchView.OnQ
         return false;
     }
 
+    @Override
+    protected void onDestroy() {
+        if (mSearchTask != null && mSearchTask.getStatus() != AsyncTask.Status.FINISHED) {
+            mSearchTask.cancel(false);
+        }
+        super.onDestroy();
+    }
+
     public void hideInputManager() {
         if (mSearchView != null) {
             if (mImm != null) {
@@ -173,6 +173,51 @@ public class SearchActivity extends BaseThemedActivity implements SearchView.OnQ
             mSearchView.clearFocus();
 
             SearchHistory.getInstance(this).addSearchString(queryString);
+        }
+    }
+
+    private class SearchTask extends AsyncTask<String,Void,ArrayList<Object>> {
+
+        @Override
+        protected ArrayList<Object> doInBackground(String... params) {
+            ArrayList<Object> results = new ArrayList<>(27);
+            List<Song> songList = SongLoader.searchSongs(SearchActivity.this, params[0], 10);
+            if (!songList.isEmpty()) {
+                results.add(getString(R.string.songs));
+                results.addAll(songList);
+            }
+
+            if (isCancelled()) {
+                return null;
+            }
+            List<Album> albumList = AlbumLoader.getAlbums(SearchActivity.this, params[0], 7);
+            if (!albumList.isEmpty()) {
+                results.add(getString(R.string.albums));
+                results.addAll(albumList);
+            }
+
+            if (isCancelled()) {
+                return null;
+            }
+            List<Artist> artistList = ArtistLoader.getArtists(SearchActivity.this, params[0], 7);
+            if (!artistList.isEmpty()) {
+                results.add(getString(R.string.artists));
+                results.addAll(artistList);
+            }
+            if (results.size() == 0) {
+                results.add(getString(R.string.nothing_found));
+            }
+            return results;
+        }
+
+        @Override
+        protected void onPostExecute(ArrayList<Object> objects) {
+            super.onPostExecute(objects);
+            mSearchTask = null;
+            if (objects != null) {
+                adapter.updateSearchResults(objects);
+                adapter.notifyDataSetChanged();
+            }
         }
     }
 }