+++ app/src/main/java/com/naman14/timber/nowplaying/Timber1.java
@@ -36,6 +36,7 @@ public class Timber1 extends BaseNowplayingFragment {
 
         setMusicStateListener();
         setSongDetails(rootView);
+        initGestures(rootView.findViewById(R.id.album_art));
 
         return rootView;
     }