+++ app/src/main/java/com/naman14/timber/utils/Constants.java
@@ -53,4 +53,17 @@ public class Constants {
     public static final String SETTINGS_STYLE_SELECTOR_WHAT = "style_selector_what";
 
     public static final String SETTINGS_STYLE_SELECTOR = "settings_style_selector";
+
+    public static final int PLAYLIST_VIEW_DEFAULT = 0;
+    public static final int PLAYLIST_VIEW_LIST = 1;
+    public static final int PLAYLIST_VIEW_GRID = 2;
+
+    public static final int PLAYLIST_ALBUM_ART_TAG = 888;
+    public static final int ACTION_DELETE_PLAYLIST = 111;
+
+
+    public static final String ACTIVITY_TRANSITION = "activity_transition";
+
+
+
 }