+++ app/src/main/java/com/naman14/timber/utils/PreferencesUtility.java
@@ -17,7 +17,7 @@ package com.naman14.timber.utils;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.content.SharedPreferences.OnSharedPreferenceChangeListener;
-import android.os.AsyncTask;
+import android.os.Environment;
 import android.preference.PreferenceManager;
 
 public final class PreferencesUtility {
@@ -33,12 +33,19 @@ public final class PreferencesUtility {
     private static final String TOGGLE_SYSTEM_ANIMATIONS = "toggle_system_animations";
     private static final String TOGGLE_ARTIST_GRID = "toggle_artist_grid";
     private static final String TOGGLE_ALBUM_GRID = "toggle_album_grid";
+    private static final String TOGGLE_PLAYLIST_VIEW = "toggle_playlist_view";
+    private static final String TOGGLE_SHOW_AUTO_PLAYLIST = "toggle_show_auto_playlist";
+    private static final String LAST_FOLDER = "last_folder";
+
     private static final String TOGGLE_HEADPHONE_PAUSE = "toggle_headphone_pause";
     private static final String THEME_PREFERNCE = "theme_preference";
     private static final String START_PAGE_INDEX = "start_page_index";
     private static final String START_PAGE_PREFERENCE_LASTOPENED = "start_page_preference_latopened";
     private static final String NOW_PLAYNG_THEME_VALUE = "now_playing_theme_value";
     private static final String TOGGLE_XPOSED_TRACKSELECTOR = "toggle_xposed_trackselector";
+    public static final String LAST_ADDED_CUTOFF = "last_added_cutoff";
+    public static final String GESTURES = "gestures";
+
     private static PreferencesUtility sInstance;
 
     private static SharedPreferences mPreferences;
@@ -72,16 +79,9 @@ public final class PreferencesUtility {
     }
 
     public void setArtistsInGrid(final boolean b) {
-        new AsyncTask<Void, Void, Void>() {
-            @Override
-            protected Void doInBackground(final Void... unused) {
-                final SharedPreferences.Editor editor = mPreferences.edit();
-                editor.putBoolean(TOGGLE_ARTIST_GRID, b);
-                editor.apply();
-                return null;
-            }
-        }.execute();
-
+        final SharedPreferences.Editor editor = mPreferences.edit();
+        editor.putBoolean(TOGGLE_ARTIST_GRID, b);
+        editor.apply();
     }
 
     public boolean isAlbumsInGrid() {
@@ -89,16 +89,9 @@ public final class PreferencesUtility {
     }
 
     public void setAlbumsInGrid(final boolean b) {
-        new AsyncTask<Void, Void, Void>() {
-            @Override
-            protected Void doInBackground(final Void... unused) {
-                final SharedPreferences.Editor editor = mPreferences.edit();
-                editor.putBoolean(TOGGLE_ALBUM_GRID, b);
-                editor.apply();
-                return null;
-            }
-        }.execute();
-
+        final SharedPreferences.Editor editor = mPreferences.edit();
+        editor.putBoolean(TOGGLE_ALBUM_GRID, b);
+        editor.apply();
     }
 
     public boolean pauseEnabledOnDetach() {
@@ -114,15 +107,9 @@ public final class PreferencesUtility {
     }
 
     public void setStartPageIndex(final int index) {
-        new AsyncTask<Void, Void, Void>() {
-            @Override
-            protected Void doInBackground(final Void... unused) {
-                final SharedPreferences.Editor editor = mPreferences.edit();
-                editor.putInt(START_PAGE_INDEX, index);
-                editor.apply();
-                return null;
-            }
-        }.execute();
+        final SharedPreferences.Editor editor = mPreferences.edit();
+        editor.putInt(START_PAGE_INDEX, index);
+        editor.apply();
     }
 
     public void setLastOpenedAsStartPagePreference(boolean preference) {
@@ -136,16 +123,9 @@ public final class PreferencesUtility {
     }
 
     private void setSortOrder(final String key, final String value) {
-        new AsyncTask<Void, Void, Void>() {
-            @Override
-            protected Void doInBackground(final Void... unused) {
-                final SharedPreferences.Editor editor = mPreferences.edit();
-                editor.putString(key, value);
-                editor.apply();
-
-                return null;
-            }
-        }.execute();
+        final SharedPreferences.Editor editor = mPreferences.edit();
+        editor.putString(key, value);
+        editor.apply();
     }
 
     public final String getArtistSortOrder() {
@@ -212,4 +192,48 @@ public final class PreferencesUtility {
     public boolean getXPosedTrackselectorEnabled() {
         return mPreferences.getBoolean(TOGGLE_XPOSED_TRACKSELECTOR, false);
     }
+
+    public int getPlaylistView() {
+        return mPreferences.getInt(TOGGLE_PLAYLIST_VIEW ,0);
+    }
+
+    public void setPlaylistView(final int i) {
+        final SharedPreferences.Editor editor = mPreferences.edit();
+        editor.putInt(TOGGLE_PLAYLIST_VIEW, i);
+        editor.apply();
+    }
+
+    public boolean showAutoPlaylist() {
+        return mPreferences.getBoolean(TOGGLE_SHOW_AUTO_PLAYLIST, true);
+    }
+
+    public void setToggleShowAutoPlaylist(final boolean b) {
+        final SharedPreferences.Editor editor = mPreferences.edit();
+        editor.putBoolean(TOGGLE_SHOW_AUTO_PLAYLIST, b);
+        editor.apply();
+    }
+
+    /** @parm lastAddedMillis timestamp in millis used as a cutoff for last added playlist */
+    public void setLastAddedCutoff(long lastAddedMillis) {
+        mPreferences.edit().putLong(LAST_ADDED_CUTOFF, lastAddedMillis).apply();
+    }
+
+    public long getLastAddedCutoff() {
+        return mPreferences.getLong(LAST_ADDED_CUTOFF, 0L);
+    }
+
+    public boolean isGesturesEnabled() {
+        return mPreferences.getBoolean(GESTURES, true);
+    }
+
+    public void storeLastFolder(String path) {
+        SharedPreferences.Editor editor = mPreferences.edit();
+        editor.putString(LAST_FOLDER, path);
+        editor.apply();
+    }
+
+    public String getLastFolder() {
+        return mPreferences.getString(LAST_FOLDER, Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_MUSIC).getPath());
+    }
 }
+